backend/app/pyproject.toml
[tool.poetry]
name = "vixenmapsapi"
version = "1.0.0"
description = "Сервис предоставляет информацию о магазинах в различных городах, с датой и временем их работы"
authors = ["darklorian <lolpokens@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
httpx = "^0.24.1"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
